# swagger/swagger.yml
info:
  title: Comentario
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
swagger: "2.0"
basePath: /api
securityDefinitions:
  # Cookie authentication for commenters
  commenterTokenHeader:
    type: apiKey
    in: header
    name: X-Commenter-Token

  # Cookie authentication for owners. Uses the apiKey type for the lack of a proper cookie-based authentication in
  # Swagger 2
  ownerCookie:
    type: apiKey
    in: header
    name: Cookie
definitions:
...
  domain:
    description: Registered domain
    type: object
    properties:
      domain:
        type: string
      ownerHex:
        $ref: "#/definitions/hexId"
      name:
        type: string
      creationDate:
        type: string
        format: date-time
      state:
        $ref: "#/definitions/domainState"
      importedComments:
        type: boolean
        x-omitempty: false
      autoSpamFilter:
        type: boolean
        x-omitempty: false
      requireModeration:
        type: boolean
        x-omitempty: false
      requireIdentification:
        type: boolean
        x-omitempty: false
      moderateAllAnonymous:
        type: boolean
        x-omitempty: false
      moderators:
        type: array
        items:
          $ref: "#/definitions/domainModerator"
      emailNotificationPolicy:
        $ref: "#/definitions/emailNotificationPolicy"
      idps:
        $ref: "#/definitions/idpMap"
      ssoSecret:
        type: string
      ssoUrl:
        type: string
        format: url
      defaultSortPolicy:
        $ref: "#/definitions/sortPolicy"
  domainModerator:
    description: Domain moderator
    type: object
    properties:
      email:
        type: string
        format: email
      domain:
        type: string
      addDate:
        type: string
        format: date-time
  domainState:
    description: Domain state
    type: string
    enum:
      - unfrozen
      - frozen
...
  emailNotificationPolicy:
    description: Email notification policy
    type: string
    enum:
      - all
      - none
      - pending-moderation
  hexId:
    description: ID consisting of 64 hex digits
    type: string
    minLength: 64
    maxLength: 64
    pattern: '[0-9a-f]{64}'
  idpMap:
    description: Map of enabled identity providers (name => boolean), including 'commento', 'sso', and all known federated IdPs
    type: object
    x-go-type:
      import:
        package: "gitlab.com/comentario/comentario/internal/api/exmodels"
      type: "IdentityProviderMap"
    x-omitempty: false
...
  sortPolicy:
    description: Sort policy
    type: string
    enum:
      - score-desc
      - creationdate-desc
      - creationdate-asc
...
paths:
...
  /domain/clear:
    post:
      operationId: DomainClear
      summary: Clear all domain's pages/comments/votes
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
              - domain
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
              domain:
                type: string
                minLength: 1
                maxLength: 253
      responses:
        204:
          description: Domain has been cleared
  /domain/delete:
    post:
      operationId: DomainDelete
      summary: Delete specified domain
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
              - domain
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
              domain:
                type: string
                minLength: 1
                maxLength: 253
      responses:
        204:
          description: Domain has been deleted
...
  /domain/list:
    post:
      operationId: DomainList
      summary: Get a list of registered domains
      security:
        - ownerCookie: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
      responses:
        200:
          description: List of domains and configured identity providers
          schema:
            type: object
            properties:
              domains:
                type: array
                items:
                  $ref: "#/definitions/domain"
              configuredOauths:
                $ref: "#/definitions/idpMap"
  /domain/moderator/delete:
    post:
      operationId: DomainModeratorDelete
      summary: Delete specified domain moderator
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
              - domain
              - email
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
              domain:
                type: string
                minLength: 1
                maxLength: 253
              email:
                type: string
                format: email
      responses:
        204:
          description: Domain moderator has been deleted
  /domain/moderator/new:
    post:
      operationId: DomainModeratorNew
      summary: Add a new domain moderator
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
              - domain
              - email
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
              domain:
                type: string
                minLength: 1
                maxLength: 253
              email:
                type: string
                format: email
      responses:
        204:
          description: Domain moderator has been added
  /domain/new:
    post:
      operationId: DomainNew
      summary: Register a new domain
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
              - name
              - domain
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
              name:
                description: Display name of the domain
                type: string
                minLength: 1
                maxLength: 63
              domain:
                type: string
                minLength: 1
                maxLength: 253 # Maximum length of a valid DNS entry
      responses:
        200:
          description: Domain added successfully
          schema:
            type: object
            properties:
              domain:
                type: string
  /domain/sso/new:
    post:
      operationId: DomainSsoSecretNew
      summary: Generate an SSO secret for specified domain
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
              - domain
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
              domain:
                type: string
                minLength: 1
                maxLength: 253
      responses:
        200:
          description: SSO secret has been generated for the domain
          schema:
            type: object
            properties:
              ssoSecret:
                $ref: "#/definitions/hexId"
  /domain/statistics:
    post:
      operationId: DomainStatistics
      summary: Get comment and view statistics for specified domain
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
              - domain
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
              domain:
                type: string
                minLength: 1
                maxLength: 253
      responses:
        200:
          description: Success or failure response (see 'success' property)
          schema:
            type: object
            properties:
              viewsLast30Days:
                type: array
                items:
                  type: integer
              commentsLast30Days:
                type: array
                items:
                  type: integer
  /domain/update:
    post:
      operationId: DomainUpdate
      summary: Update properties of specified domain
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - ownerToken
              - domain
            properties:
              ownerToken:
                $ref: "#/definitions/hexId"
              domain:
                $ref: "#/definitions/domain"
      responses:
        204:
          description: Domain properties have been updated

  #---------------------------------------------------------------------------------------------------------------------
  # Emails
  #---------------------------------------------------------------------------------------------------------------------
...